헐 병합을 안했었나?
병합을 안했네
