+++ src/main/webapp/resources/css/GwCss.css
@@ -953,7 +953,7 @@ li.nav-folder-close .edit-btn {top:10px;right:15px}
 - (프로젝트 우측 모아보기 메뉴)
 ------------------------------------------*/
 ul.pro-gather-nav {display:block;width:100%;height:60px;margin:0}
-ul.pro-gather-nav li {float:left;width:20%}
+ul.pro-gather-nav li {float:left;width:25%}
 ul.pro-gather-nav li a {display:block;width:100%;height:60px;padding-top:12px;text-align:center}
 ul.pro-gather-nav li a:hover {background-color:#f1f2f3}
